requirements/common.txt
Django==3.2.19
psycopg2-binary==2.9.6
django-dynamic-raw-id==3.0
